Adding PS4
Problem set 4 focused on recursion and on encrypting a decrypting text.

The first part of the problem set was to create a function that generated all unique permutations of a string. This had to be implemented via a recursive method.

The second part was to encrypt a piece of text via a Julian cipher. Then decrypt it based on what produced the most valid words.

Finally, given a permutation of the vowels had to encrypt a piece of text so that the vowels were exchanged as determined by the given permutation, all consonants remained unchanged. Then had to decrypt the message.