Adding PS5
This problem set involved making Triggers to alert if news stories from Google or Yahoo were relevant.

The triggers were made from inherited classes that built upon each other.

The code I have written passed all the unit tests built for this problem set. However, it failed to run as once completed. It appeared to be an issue with the objects getting a description. This seems to be an error within the main code provided - potentially due to running out of date code - as have found other people online getting the same issue.

Whilst it may provide useful experience to go through and debug the code fully for now I'm going to move on and potentially come back to this later if I get the chance.