Completed PS2 minus written proof.
This commit contains a completed problem set 2 with all the theory questions answered, except from the written proof for why the 3D rendering is O(1). In short this is because, like part a, each level (i) in the tree contributes to the complexity but at each level the cost decreases proportional to the area so that complexity is (3*4^i) *((1/9)^i). This can - and will - be shown to have a constant bound.

The second part of this problem set is the practical side and have now implemented a working min heap data structure that solves all the test cases in an acceptable amount of time.