Changed implementation of Sweep-line Algorithm
After re-running the cProfile again, on my previous upload, intersections was still the most expensive function used by the CPU. Based on the follow up questions this clearly wasn't right as intersection isn't a data type.

So after going through the code given noted that intersections was a superfluous function, as the range_index contained all the wires that needed to be checked. This meant only need to check if the wires in the range_index are within the the y values of the vertical lines. The pre-built function count did exactly that.

So have changed the implementation and count is now the most expensive part of the program for the cpu. This makes sense with the problem sheet as count looks through an array and can improve this based by making the array an AVL tree that has the ranking features discussed earlier in the problem set.